pdok-1263: Code review and refactor
