+++ src/main/java/nl/pdok/catalog/Catalog.java
@@ -13,7 +13,7 @@ import nl.pdok.catalog.extract.ExtractConfiguration;
 import nl.pdok.catalog.featured.FeatureTemplate;
 import nl.pdok.catalog.featured.FeaturedCollectionOptions;
 import nl.pdok.catalog.job.JobConfigurationDataset;
-import nl.pdok.catalog.jsonentities.JobEntry;
+import nl.pdok.catalog.jobentry.JobEntry;
 import nl.pdok.catalog.testdata.TestData;
 import nl.pdok.catalog.tiling.TilingConfiguration;
 import nl.pdok.catalog.workbench.Workbench;
