+++ src/main/java/nl/pdok/catalog/FileSystemCatalog.java
@@ -28,12 +28,12 @@ import nl.pdok.catalog.extract.ExtractConfiguration;
 import nl.pdok.catalog.extract.ExtractConfigurationReader;
 import nl.pdok.catalog.featured.FeatureTemplate;
 import nl.pdok.catalog.featured.FeaturedCollectionOptions;
-import nl.pdok.catalog.job.GitInteractionsHandler;
+import nl.pdok.catalog.gitutil.GitInteractionsHandler;
 import nl.pdok.catalog.job.JobConfiguration;
 import nl.pdok.catalog.job.JobConfigurationDataset;
 import nl.pdok.catalog.job.JobConfigurationReader;
-import nl.pdok.catalog.job.JobEntriesReader;
-import nl.pdok.catalog.jsonentities.JobEntry;
+import nl.pdok.catalog.jobentry.JobEntriesReader;
+import nl.pdok.catalog.jobentry.JobEntry;
 import nl.pdok.catalog.testdata.TestData;
 import nl.pdok.catalog.tiling.TilingConfiguration;
 import nl.pdok.catalog.tiling.TilingConfigurationReader;
