+++ src/test/java/nl/pdok/catalog/FileSystemCatalogTest.java
@@ -14,7 +14,7 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.Set;
 import nl.pdok.catalog.featured.FeatureTemplate;
-import nl.pdok.catalog.jsonentities.JobEntry;
+import nl.pdok.catalog.jobentry.JobEntry;
 import nl.pdok.catalog.workbench.FmeWorkbenchEnvConfig;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Assert;
